Factor out Kore.Log.KoreLogOptions
